Refactor: Seperated business logic from component
